Refactor folder structure and enable transpilation to es6 and browser before publishing the package
