+++ package.json
@@ -7,6 +7,9 @@
     "test": "tests"
   },
   "scripts": {
+    "compile-es5": "./node_modules/.bin/babel -d lib/ src/",
+    "compile-browser": "./node_modules/.bin/browserify lib/contextjs.js --outfile lib/browser-contextjs.js --standalone cop",
+    "prepublish": "npm run compile-es5 && npm run compile-browser",
     "test": "npm run test-node && npm run test-karma",
     "test-karma": "karma start --single-run",
     "karma": "karma",
@@ -44,6 +47,7 @@
     "babel-plugin-array-includes": "^2.0.3",
     "babel-polyfill": "^6.8.0",
     "babel-preset-es2015": "^6.6.0",
+    "browserify": "^13.0.1",
     "chai": "^3.5.0",
     "karma": "^0.13.22",
     "karma-babel-preprocessor": "^6.0.1",
